refactor: `NicknameUpdateError.toUiText` 확장 함수를 별도 매퍼 파일로 분리
