+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/data/repository/member/NicknameUpdateError.kt
@@ -1,16 +1,6 @@
 package com.yagubogu.data.repository.member
 
 import com.yagubogu.data.util.ApiException
-import com.yagubogu.ui.util.UiText
-import yagubogu.composeapp.generated.resources.Res
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_duplicate
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_invalid_format
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_member_not_found
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_network_error
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_no_permission
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_server_error
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_too_long
-import yagubogu.composeapp.generated.resources.setting_edit_nickname_unknown_error
 
 sealed interface NicknameUpdateError {
     data object DuplicateNickname : NicknameUpdateError // 409: 이미 존재함
@@ -43,18 +33,4 @@ fun Throwable.toNicknameUpdateError(): NicknameUpdateError =
         is ApiException.ServerError -> NicknameUpdateError.ServerError // 5xx
         is ApiException.NetworkError -> NicknameUpdateError.NetworkIssue // 네트워크 미연결
         else -> NicknameUpdateError.Unknown(message)
-    }
-
-fun NicknameUpdateError.toUiText(): UiText =
-    when (this) {
-        NicknameUpdateError.DuplicateNickname -> UiText.StringRes(Res.string.setting_edit_nickname_duplicate)
-        NicknameUpdateError.InvalidNickname -> UiText.StringRes(Res.string.setting_edit_nickname_invalid_format)
-        NicknameUpdateError.MemberNotFound -> UiText.StringRes(Res.string.setting_edit_nickname_member_not_found)
-        NicknameUpdateError.NoPermission -> UiText.StringRes(Res.string.setting_edit_nickname_no_permission)
-        NicknameUpdateError.PayloadTooLarge -> UiText.StringRes(Res.string.setting_edit_nickname_too_long)
-        NicknameUpdateError.ServerError -> UiText.StringRes(Res.string.setting_edit_nickname_server_error)
-        NicknameUpdateError.NetworkIssue -> UiText.StringRes(Res.string.setting_edit_nickname_network_error)
-        is NicknameUpdateError.Unknown ->
-            message?.let { UiText.DynamicString(it) }
-                ?: UiText.StringRes(Res.string.setting_edit_nickname_unknown_error)
     }
