+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/setting/SettingViewModel.kt
@@ -7,8 +7,8 @@ import com.yagubogu.data.repository.auth.AuthRepository
 import com.yagubogu.data.repository.member.MemberRepository
 import com.yagubogu.data.repository.member.NicknameUpdateError
 import com.yagubogu.data.repository.member.toNicknameUpdateError
-import com.yagubogu.data.repository.member.toUiText
 import com.yagubogu.data.repository.thirdparty.ThirdPartyRepository
+import com.yagubogu.ui.mapper.text.toUiText
 import com.yagubogu.ui.mapper.toUiModel
 import com.yagubogu.ui.setting.model.MemberInfoItem
 import com.yagubogu.ui.setting.model.PresignedUrlCompleteItem
